Move encoding into own file
