+++ phe/__init__.py
@@ -1,10 +1,11 @@
 import pkg_resources
 
+from phe.encoding import EncodedNumber
 from phe.paillier import generate_paillier_keypair
-from phe.paillier import EncodedNumber
 from phe.paillier import EncryptedNumber
 from phe.paillier import PaillierPrivateKey, PaillierPublicKey
 from phe.paillier import PaillierPrivateKeyring
+
 import phe.util
 
 try:
